# backend/specs/openapi.yaml
openapi: 3.1.0
info:
  title: NarrativeOS API
  version: 0.3.0
paths:
...
  /v1/reader/snapshot:
    get:
      tags:
      - product-runtime
      summary: Reader Snapshot
      operationId: reader_snapshot_v1_reader_snapshot_get
      parameters:
      - name: session_id
        in: query
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Reader Snapshot V1 Reader Snapshot Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - product-runtime
      summary: Reader Snapshot Post
      operationId: reader_snapshot_post_v1_reader_snapshot_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReaderSnapshotRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Reader Snapshot Post V1 Reader Snapshot Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/scene/advance:
    post:
      tags:
      - product-runtime
      summary: Scene Advance
      operationId: scene_advance_v1_scene_advance_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SceneAdvanceRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Scene Advance V1 Scene Advance Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/timeline/worldlines/{worldline_id}/loom:
    get:
      tags:
      - product-runtime
      summary: Worldline Runtime Events
      operationId: worldline_runtime_events_v1_timeline_worldlines__worldline_id__loom_get
      parameters:
      - name: worldline_id
        in: path
        required: true
        schema:
          type: string
          title: Worldline Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Worldline Runtime Events V1 Timeline Worldlines  Worldline
                  Id  Loom Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/quality/evaluate:
    post:
      tags:
      - product-runtime
      summary: Quality Evaluate
      operationId: quality_evaluate_v1_quality_evaluate_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QualityEvaluateRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Quality Evaluate V1 Quality Evaluate Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/canon/commit:
    post:
      tags:
      - product-runtime
      summary: Canon Commit
      operationId: canon_commit_v1_canon_commit_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CanonCommitRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Canon Commit V1 Canon Commit Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CanonCommitRequest:
      properties:
        candidate_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Candidate Id
        session_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Id
        world_id:
          anyOf:
          - type: string
          - type: 'null'
          title: World Id
        worldline_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Worldline Id
        world_version_id:
          anyOf:
          - type: string
          - type: 'null'
          title: World Version Id
        chapter_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Chapter Id
        target_status:
          type: string
          title: Target Status
          default: canon
        confirmed:
          type: boolean
          title: Confirmed
          default: false
        confirmed_by:
          anyOf:
          - type: string
          - type: 'null'
          title: Confirmed By
        quality_report:
          additionalProperties: true
          type: object
          title: Quality Report
      type: object
      title: CanonCommitRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    QualityEvaluateRequest:
      properties:
        body:
          type: string
          title: Body
        candidate_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Candidate Id
        session_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Id
        world_id:
          anyOf:
          - type: string
          - type: 'null'
          title: World Id
        world_version_id:
          anyOf:
          - type: string
          - type: 'null'
          title: World Version Id
        choices:
          items:
            type: string
          type: array
          title: Choices
        character_fidelity_score:
          type: number
          title: Character Fidelity Score
          default: 0.6
        ending_ready:
          type: boolean
          title: Ending Ready
          default: false
        paywall_required:
          type: boolean
          title: Paywall Required
          default: false
      type: object
      required:
      - body
      title: QualityEvaluateRequest
...
    ReaderSnapshotRequest:
      properties:
        session_id:
          type: string
          title: Session Id
      type: object
      required:
      - session_id
      title: ReaderSnapshotRequest
...
    SceneAdvanceRequest:
      properties:
        session_id:
          type: string
          title: Session Id
        choice_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Choice Id
        freeform_intent:
          anyOf:
          - type: string
          - type: 'null'
          title: Freeform Intent
        account_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Account Id
        reader_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Reader Id
        worldline_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Worldline Id
        scene_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Scene Id
        user_id:
          anyOf:
          - type: string
          - type: 'null'
          title: User Id
      type: object
      required:
      - session_id
      title: SceneAdvanceRequest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...